refactor: pub-sub 이름 변경
- pub-sub 이름을 broker로 변경
- play, chat 모듈에 변경사항 반영
- 디렉터리 위치 변경
