+++ apps/backend/src/core/broker/interfaces/message.interface.ts
